Separated code into multiple components
