+++ components/Header.js
@@ -1,0 +1,7 @@
+
+export default function Header() {
+  return (
+    <header className='py-2 w-full h-fit bg-gray-800 flex flex-row flex-wrap items-center text-center justify-center gap-4'>
+    </header>
+  )
+}
